+++ src/js/helpers.js
@@ -1,4 +1,7 @@
-
+// Returns a function, that, as long as it continues to be invoked, will not
+// be triggered. The function will be called after it stops being called for
+// N milliseconds. If `immediate` is passed, trigger the function on the
+// leading edge, instead of the trailing.
 export function debounce(func, wait, immediate) {
 	var timeout;
 	return function() {
